+++ app/(routes)/product/[id]/components/index.ts
@@ -1,0 +1,3 @@
+export { default as BackIconButton } from './BackIconButton';
+export { default as ProductContents } from './ProductContents';
+export { default as ProductOverview } from './ProductOverview';
