+++ app/components/common/GlobalNavigationBar.tsx
@@ -1,10 +1,10 @@
 'use client';
 
+import clsx from 'clsx';
+import { useRouter } from 'next/navigation';
 import { Box, Divider, Typography } from '@mui/material';
 import { PRODUCT_MENU } from '@/app/modules/constants';
 import { productStore, IProductState } from '@/app/modules/store/productStore';
-import { useRouter } from 'next/navigation';
-import clsx from 'clsx';
 
 /**
  * 레이아웃 > 상단 네비게이션 바 컴포넌트
