+++ app/components/index.ts
@@ -1,18 +1,14 @@
 export { default as FloatingUpButton } from './common/FloatingUpButton';
 export { default as Header } from './common/Header';
-export { default as SkeletonUI } from './common/SkeletonUI';
-export { default as BackIconButton } from './common/BackIconButton';
-
-export { default as FilterChip } from './FilterChip/FilterChip';
-export { default as FilterChipsBox } from './FilterChip/FilterChipsBox';
+export { default as FilterChip } from './common/FilterChip/FilterChip';
+export { default as FilterChipsBox } from './common/FilterChip/FilterChipsBox';
+export { default as GlobalNavigationBar } from './common/GlobalNavigationBar';
+export { default as SNBFilterList } from './common/SideNavigationBar/SNBFilterList';
+export { default as SideNavigationBar } from './common/SideNavigationBar/SideNavigationBar';
 
 export { default as ProductColor } from './Product/ProductColor';
-export { default as ProductContents } from './Product/ProductContents';
 export { default as ProductInfo } from './Product/ProductInfo';
-export { default as ProductOverview } from './Product/ProductOverview';
 
-export { default as SideNavigationBarList } from './SideNavigationBar/SideBarList';
-export { default as SideNavigationBar } from './SideNavigationBar/SideNavigationBar';
+export { default as DisplayProducts } from './DisplayProducts';
 
-export { default as DisplayProducts } from './common/DisplayProducts';
-export { default as GlobalNavigationBar } from './common/GlobalNavigationBar';
+export { default as SkeletonUI } from './SkeletonUI';
